Implemented review comments
