+++ .gitignore
@@ -1,5 +1,5 @@
 .DS_Store
-**/.build/**
+.build
 /Packages
 xcuserdata/
 DerivedData/
